[CI] Let git-numstat own the diff invocation, not just parsing
Review follow-up. The numstat parser assumed the caller passed -z, an
indirect dependency on the runner's behavior. Move the `git diff --numstat`
invocation into the module so the flags that make the output parseable
(-z, --no-renames) and the maxBuffer live next to the code that relies on
them, and rename the module (parse-git-numstat -> git-numstat) since it now
does more than parse.

- git-numstat.mjs exports changedFileStats(baseSha, head), which runs the
  diff and returns [{ path, lines }]. parseGitNumstat stays exported for the
  unit tests.
- label-pr.mjs just calls changedFileStats(); it no longer knows the diff
  flags or the buffer size.

No behavior change.
